Satisfy iOS viewport policy gates
