+++ Packages/iOS/CmuxMobileTerminal/Sources/CmuxMobileTerminal/GhosttyRuntime.swift
@@ -384,7 +384,14 @@ public final class GhosttyRuntime {
         }
 
         #if DEBUG
-        if handleBottomScrollStressDebugAction(action, target: target) {
+        if action.tag == GHOSTTY_ACTION_SCROLLBAR {
+            let sb = action.action.scrollbar
+            MobileDebugLog.anchormux("scroll.bar total=\(sb.total) offset=\(sb.offset) len=\(sb.len)")
+            if target.tag == GHOSTTY_TARGET_SURFACE, let surface = target.target.surface {
+                Task { @MainActor in
+                    GhosttySurfaceView.view(for: surface)?.recordBottomScrollStressScrollbar(total: Int(sb.total), offset: Int(sb.offset), len: Int(sb.len))
+                }
+            }
             return true
         }
         #endif
